Fixes for compatibility with Elixir 1.5 / OTP 20.0
* Upgrades excheck to 0.5.3

This upgrade is necessary because as of Elixir 
1.4, all ExUnit CLI formatters must use the 
GenServer API (instead of the former GenEvent API)
. Newer versions of this lib account for this 
change.

* Agnostic parsing of content with escaped values

:xmerl_scan parses inner text with escaped values
(like ampersand) in different ways depending on 
the version of the runtime. In older versions, 
the parser splits on the escaped value. This
commit refactors the test to turn this part of the
test agnostic.